Refactor multipurpose utility serializer functions to its own file
